+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 **NOTE**: As permitted by semantic versioning, backward compatibility is NOT maintained for initial development, i.e. releases before 1.0.0.
 
 ## Unreleased
-Nothing yet.
+### Fixed
+- Newlines are no longer semi-randomly printed to standard error in preprocessing and assembling mode.
 
 ## [0.7.0] - 2019-02-13
 ### Added