+++ CHANGELOG.md
@@ -5,13 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 **NOTE**: As permitted by semantic versioning, backward compatibility is NOT maintained for initial development, i.e. releases before 1.0.0.
 
-## Unreleased
-### Fixed
-- Newlines are no longer semi-randomly printed to standard error in preprocessing and assembling mode.
 
+## Unreleased
+Nothing yet!
+
+## [0.8.0] - 2019-03-07
 ### Added
+- The disassembler subcommand.
 - The `--throttle` flag, to limit the number of operations executed on a single run.
 - The `--credits` flag.
+
+### Fixed
+- Newlines are no longer semi-randomly printed to standard error in preprocessing and assembling mode.
 
 ## [0.7.0] - 2019-02-13
 ### Added